Fixed issues with settings toolbar background on scroll. Implemented tabs in source control settings page.
