+++ CodeEdit/Features/Settings/Pages/SourceControlSettings/SourceControlGitView.swift
@@ -8,63 +8,37 @@
 import SwiftUI
 
 struct SourceControlGitView: View {
-
-    // MARK: - View
-
-    var body: some View {
-        Group {
-            Section("Git") {
-                authorNameAndEmail
-            }
-            Section {
-                ignoredFiles
-            }
-            Section {
-                options
-            }
-        }
-//        .formStyle(.grouped)
-    }
-
     @StateObject
     private var prefs: AppPreferencesModel = .shared
+
+    var body: some View {
+        SettingsForm {
+            Section {
+                gitAuthorName
+                gitEmail
+            }
+            Section {
+                ignoredFilesText
+            }
+            Section {
+                preferToRebaseWhenPulling
+                showMergeCommitsInPerFileLog
+            }
+        }
+    }
 }
 
 private extension SourceControlGitView {
-
-    // MARK: - Sections
-
-    @ViewBuilder
-    private var authorNameAndEmail: some View {
-        gitAuthorName
-        gitEmail
-    }
-
-    private var ignoredFiles: some View {
-        ignoredFilesText
-    }
-
-    @ViewBuilder
-    private var options: some View {
-        preferToRebaseWhenPulling
-        showMergeCommitsInPerFileLog
-    }
-
-    // MARK: - Preference Views
-
     private var gitAuthorName: some View {
-        TextField("Git Author Name", text: $prefs.preferences.sourceControl.git.authorName)
+        TextField("Author Name", text: $prefs.preferences.sourceControl.git.authorName)
     }
 
     private var gitEmail: some View {
-        TextField("Git Email", text: $prefs.preferences.sourceControl.git.authorEmail)
+        TextField("Author Email", text: $prefs.preferences.sourceControl.git.authorEmail)
     }
 
-    @ViewBuilder
     private var ignoredFilesText: some View {
-        HStack {
-            Text("Ignored files")
-            Spacer()
+        LabeledContent("Ignored files") {
             Button("Select files...") {
                 // TODO: Implement a view (that works) for this
             }
