+++ CodeEdit/Features/Settings/Pages/SourceControlSettings/SourceControlSettingsView.swift
@@ -8,10 +8,29 @@
 import SwiftUI
 
 struct SourceControlSettingsView: View {
+    @State var selectedTab: String = "general"
+
     var body: some View {
-        SettingsForm {
-            SourceControlGeneralView()
-            SourceControlGitView()
+        Group {
+            switch selectedTab {
+            case "general":
+                SourceControlGeneralView()
+            case "git":
+                SourceControlGitView()
+            default:
+                SourceControlGeneralView()
+            }
+        }
+        .safeAreaInset(edge: .top, spacing: 0) {
+            Picker("", selection: $selectedTab) {
+                Text("General").tag("general")
+                Text("Git").tag("git")
+            }
+            .pickerStyle(.segmented)
+            .labelsHidden()
+            .padding(.horizontal, 20)
+            .padding(.bottom, 20)
+
         }
     }
 }
